Add: Utilities.h - Added Utilities.h for handling utility functions, e.g. making and comparing files - Refactor: moved getPostScriptHeader and makePostScriptFile to Utilities
