+++ include/Shape.hpp
@@ -9,7 +9,6 @@
 #include <string>
 #include <memory>
 #include <iostream>
-#include <fstream>
 
 // *********************************************************************
 // Point Type
@@ -57,13 +56,5 @@ protected:
 	virtual std::string generatePostScript(point_t center) const = 0;
 
 };
-
-// *********************************************************************
-// Utility Functions
-// *********************************************************************
-
-std::string getPostScriptHeader();
-
-void makePostScriptFile(const Shape & shape, std::string filename);
 
 #endif // #ifndef FILE_SHAPE_HPP_INCLUDED
